ajuste na validação da JOB

+++ api/jobs/resetApp.js
@@ -3,7 +3,7 @@ const User = require('../models/User');
 const Message = require('../models/Message');
 const bcrypt = require('bcryptjs');
 const dotenv = require('dotenv');
-const fs = require('fs.Promises');
+const fs = require('fs').promises;
 const path =  require('path');
 
 dotenv.config();
@@ -42,9 +42,11 @@ schedule.scheduleJob('0 6 */1 * *' , async function () {
     try {
         const folderPath = path.join(__dirname, '../uploads')
         const files = await fs.readdir(folderPath);
-        for (const file of files) {
-            await fs.unlink(path.resolve(folderPath, file));
-            console.log(`${folderPath}/${file} has been removed successfully`);
+        if(files.length > 0) {
+            for (const file of files) {
+                await fs.unlink(path.resolve(folderPath, file));
+                console.log(`${folderPath}/${file} has been removed successfully`);
+            }
         }
     } catch (err){
         console.log(err);
